+++ lib/src/blocs/page/bloc.dart
@@ -2,6 +2,7 @@ import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 
 import '../../models/back_pressed_result_enum.dart';
+import '../../pages/abstract.dart';
 import '../../widgets/navigator.dart';
 import '../page_stack/bloc.dart';
 import 'close_event.dart';
@@ -12,7 +13,8 @@ import 'event.dart';
 /// A BLoC that backs each stateful page of your app.
 ///
 /// [C] is the base class for all app's page configurations.
-class PageBloc<C extends PageConfiguration> {
+/// [R] is the result returned when the page pops.
+class PageBloc<C extends PageConfiguration, R> {
   final _eventsController = BehaviorSubject<PageBlocEvent>();
 
   Stream<PageBlocEvent> get events => _eventsController.stream;
@@ -40,19 +42,39 @@ class PageBloc<C extends PageConfiguration> {
   }
 
   /// Emits [PageBlocCloseEvent] for [PageStackBloc] to remove and dispose
+  /// the current page. [data] is passed to the new topmost [PageBloc] and
+  /// can be used as the modal dialog result.
+  void pop([R? data]) {
+    _eventsController.sink.add(
+      data == null
+          ? const PageBlocCloseEvent(data: null)
+          : PageBlocCloseEvent<R>(data: data),
+    );
+  }
+
+  /// Emits [PageBlocCloseEvent] for [PageStackBloc] to remove and dispose
   /// the current page.
+  @Deprecated('Use pop')
+  @nonVirtual
   void closeScreen() {
-    _eventsController.sink.add(const PageBlocCloseEvent());
+    _eventsController.sink.add(const PageBlocCloseEvent(data: null));
   }
 
   /// Emits [event] for [PageStackBloc] to remove and dispose the current page.
   /// [event] is passed to the new topmost [PageBloc] and can be used
   /// as modal dialog result.
+  @Deprecated('Use pop for PageBlocCloseEvent or emitEvent for custom events')
+  @nonVirtual
   void closeScreenWith(PageBlocCloseEvent event) {
     _eventsController.sink.add(event);
   }
 
   /// Override this to use the result of a closed modal screen.
+  void didPopNext(AbstractPage page, PageBlocCloseEvent event) {}
+
+  /// Override this to use the result of a closed modal screen.
+  @Deprecated('Use didPopNext')
+  @nonVirtual
   void onForegroundClosed(PageBlocCloseEvent event) {}
 
   /// Called when Android back button is pressed with this page active.