+++ lib/src/blocs/page/close_event.dart
@@ -5,6 +5,10 @@ import 'event.dart';
 ///
 /// The page could be anywhere in the stack as long as at least one other page
 /// is left in the stack afterwards.
-class PageBlocCloseEvent extends PageBlocEvent {
-  const PageBlocCloseEvent();
+class PageBlocCloseEvent<R> extends PageBlocEvent {
+  final R? data;
+
+  const PageBlocCloseEvent({
+    required this.data,
+  });
 }