+++ lib/src/pages/stateless_material.dart
@@ -3,14 +3,14 @@ import 'package:flutter/widgets.dart';
 import '../blocs/page/configuration.dart';
 import 'material.dart';
 
-abstract class StatelessMaterialPage<C extends PageConfiguration>
-    extends AbstractMaterialPage<C> {
+abstract class StatelessMaterialPage<C extends PageConfiguration, R>
+    extends AbstractMaterialPage<C, R> {
   final C? configuration;
 
   @override
   C? getConfiguration() => configuration;
 
-  const StatelessMaterialPage({
+  StatelessMaterialPage({
     required ValueKey<String> super.key,
     required super.child,
     super.factoryKey,