+++ lib/src/widgets/stateful_bloc_widget.dart
@@ -4,7 +4,8 @@ import '../blocs/page/configuration.dart';
 import '../blocs/page/stateful_bloc.dart';
 
 abstract class StatefulBlocWidget<
-    B extends PageStatefulBloc<PageConfiguration, S>,
+    R,
+    B extends PageStatefulBloc<PageConfiguration, S, R>,
     S> extends StatelessWidget {
   final B bloc;